+++ core/src/main/java/com/graphhopper/routing/util/FerrySpeedCalculator.java
@@ -5,7 +5,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class FerrySpeedCalculator {
-
     private static final Logger LOGGER = LoggerFactory.getLogger(FerrySpeedCalculator.class);
     private final double unknownSpeed, minSpeed, maxSpeed;
 
@@ -15,47 +14,37 @@ public class FerrySpeedCalculator {
         this.unknownSpeed = unknownSpeed;
     }
 
-    /**
-     * Special handling for ferry ways.
-     */
     public double getSpeed(ReaderWay way) {
-        // todo: we should re-consider whether we should deal with ferries by determining (only) a speed at all.
-        //       for example this way we cannot account for waiting times for short ferries (speed is too low) or
-        //       'fast' ferries that exceed the max speed for e.g. the foot encoder. Maybe we could add an additional
-        //       encoded value that stores the ferry waiting time (maybe just short/medium/long or so) and add a few
-        //       more possible speed values to account for high ferry speeds for the slow vehicles like foot.
+        // todo: We currently face twe problems related to ferry speeds:
+        //       1) We cannot account for waiting times for short ferries (makes the ferry speed slower than we can store)
+        //       2) When the ferry speed is larger than the maximum speed of the encoder (like 15km/h for foot) the
+        //          ferry speed will be faster than we can store.
+        //       Maybe we could add an additional encoded value that stores the ferry waiting time
+        //       (maybe just short/medium/long or so) and add a few more possible speed values to account for high ferry
+        //       speeds for the slow vehicles like foot.
 
-        // During the reader process we have added the artificial "edge_distance" tag and converted the duration tag to
-        // an artificial tag called "duration:seconds"
-        double distanceInKm = way.getTag("edge_distance", -1.0) / 1000;
-        if (distanceInKm < 0)
-            throw new IllegalStateException("The artificial 'edge_distance' tag is missing for way: " + way.getId());
-        Long duration = way.getTag("duration:seconds", 0L);
-        double durationInHours = duration / 60d / 60d;
-        if (durationInHours > 0) {
-            // If duration is available we can calculate the speed. We make it slower by a factor of 1.4 to account
-            // for waiting time.
-            double calculatedTripSpeed = distanceInKm / durationInHours / 1.4;
-            // Plausibility check especially for the case of wrongly used PxM format with the intention to
-            // specify the duration in minutes, but actually using months
-            if (calculatedTripSpeed > 0.01d) {
-                if (calculatedTripSpeed > maxSpeed)
-                    return maxSpeed;
-                // If the speed is lower than the speed we can store, we have to set it to the minSpeed, but > 0
-                if (Math.round(calculatedTripSpeed) < minSpeed)
-                    return minSpeed;
-                return Math.round(calculatedTripSpeed);
-            } else {
-                LOGGER.warn("Unrealistic long duration ignored in way with way ID=" + way.getId() + " : Duration tag value="
-                        + way.getTag("duration") + " (=" + Math.round(duration / 60d) + " minutes), distance=" + distanceInKm + "km");
+        // OSMReader adds the artificial 'speed_from_duration', 'duration:seconds' and 'way_distance' tags that we can
+        // use to set the ferry speed. Otherwise we need to use fallback values.
+        double speedInKmPerHour = way.getTag("speed_from_duration", Double.NaN);
+        if (!Double.isNaN(speedInKmPerHour)) {
+            // we reduce the speed to account for waiting time (we increase the duration by 40%)
+            double speedWithWaitingTime = speedInKmPerHour / 1.4;
+            return Math.round(Math.max(minSpeed, Math.min(speedWithWaitingTime, maxSpeed)));
+        } else {
+            // the speed cannot be derived from the duration tag, we have to take a guess based on the distance.
+            double wayDistance = way.getTag("way_distance", Double.NaN);
+            if (Double.isNaN(wayDistance))
+                // we don't know the distance of this way either. probably a broken way at the border of the map.
+                return unknownSpeed;
+            else if (wayDistance < 300)
+                // use the slowest possible speed for very short ferries
+                return minSpeed;
+            else {
+                LOGGER.warn("Long ferry OSM way without valid duration: " + way.getId() + ", distance: " + wayDistance);
+                // todonow: use the unknown speed for now, but check how many such ferries actually exist and maybe use a
+                //          faster speed for longer ones
+                return unknownSpeed;
             }
         }
-
-        // duration was not present or calculated trip speed was too small
-        if (distanceInKm <= 0.3)
-            return minSpeed;
-        else
-            // unknown speed -> put penalty on ferry transport
-            return unknownSpeed;
     }
 }